+++ src/Oro/Bundle/ImapBundle/Controller/MicrosoftAccessTokenController.php
@@ -2,6 +2,8 @@
 
 namespace Oro\Bundle\ImapBundle\Controller;
 
+use Oro\Bundle\ImapBundle\Provider\MicrosoftOAuthProvider;
+use Oro\Bundle\ImapBundle\Provider\MicrosoftOAuthScopeProvider;
 use Oro\Bundle\ImapBundle\Provider\OAuthProviderInterface;
 use Oro\Bundle\ImapBundle\Provider\OAuthScopeProviderInterface;
 use Symfony\Component\HttpFoundation\JsonResponse;
@@ -33,7 +35,7 @@ class MicrosoftAccessTokenController extends AbstractAccessTokenController
      */
     protected function getOAuthProvider(): OAuthProviderInterface
     {
-        return $this->get('oro_imap.microsoft_oauth_provider');
+        return $this->get(MicrosoftOAuthProvider::class);
     }
 
     /**
@@ -62,7 +64,7 @@ class MicrosoftAccessTokenController extends AbstractAccessTokenController
 
     private function getOAuthScopeProvider(): OAuthScopeProviderInterface
     {
-        return $this->get('oro_imap.microsoft_oauth_scope_provider');
+        return $this->get(MicrosoftOAuthScopeProvider::class);
     }
 
     private function storeResponse(SessionInterface $session, Response $response): Response
@@ -106,5 +108,16 @@ class MicrosoftAccessTokenController extends AbstractAccessTokenController
         $response['email_address'] = $userInfo->getEmail();
 
         return new JsonResponse($response);
+    }
+
+    public static function getSubscribedServices()
+    {
+        return array_merge(
+            parent::getSubscribedServices(),
+            [
+                MicrosoftOAuthProvider::class,
+                MicrosoftOAuthScopeProvider::class,
+            ]
+        );
     }
 }