+++ src/Oro/Bundle/WorkflowBundle/Migrations/Data/ORM/UpdateAccessLevelsOfTransitionsPermissions.php
@@ -6,6 +6,7 @@ use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\Common\DataFixtures\DependentFixtureInterface;
 use Doctrine\Persistence\ObjectManager;
+use Oro\Bundle\DistributionBundle\Handler\ApplicationState;
 use Oro\Bundle\FrontendBundle\Migrations\Data\ORM\LoadUserRolesData;
 use Oro\Bundle\SecurityBundle\Acl\Extension\AclExtensionInterface;
 use Oro\Bundle\SecurityBundle\Acl\Extension\ObjectIdentityHelper;
@@ -37,7 +38,7 @@ class UpdateAccessLevelsOfTransitionsPermissions extends AbstractUpdatePermissio
      */
     public function load(ObjectManager $manager): void
     {
-        if (!$this->container->hasParameter('installed') || !$this->container->getParameter('installed')) {
+        if (!$this->container->get(ApplicationState::class)->isInstalled()) {
             return;
         }